✨ feat: Add SolidStudio's new open source JSON Schema Generator for OCPI
SolidStudio has released a new open source [JSON Schema Generator for OCPI](https://github.com/solidstudiosh/ocpi-schema). This generator allows developers to easily create a set of files in JSON-Schema format that reflect the communication interfaces for OCPI versions 2.1.1, 2.2, and 2.2.1.

The SolidStudio JSON Schema Generator for OCPI provides valuable API validation and assessment capabilities, ensuring compliance with the OCPI documentation. It is a user-friendly tool that can save developers time and effort. If you are working on an OCPI API, I encourage you to check out this generator.