✨ feat: Add version introduction documentation
Added a new file "01-version-intro.md" under the "06-versions" directory. The file includes introductory information about versions in OCPI. The module identifier is "versions" and it is classified as a configuration module. This module serves as the required base module for OCPI and provides information about supported versions and modules through OCPI connections.

The file includes details on how clients can determine the supported versions of OCPI and the corresponding modules using the provided links.

Fixes #123